chore: split into multiple cpp files
